Rewrite the README to focus on JSON.
Rewrote the introduction to explain VarioML a bit more in detail, and to
focus on the JSON implementation. Provided more information on the JSON
implementation.

Fixed more broken links, although also created a link to the new JSON
schema file that hasn't been created yet.

Reshuffled the README a bit.