Rename Source to DataSource to prevent clashes.
The original VarioML reuses the term "source" for three different
things. It was used for a genetic source (now renamed to
"genetic_source" in JSON Schema), for a data source, and for providing a
source to an accession. Now renaming the Source object to DataSource and
renaming its mentions to "data_source" to prevent problems with clashes
with the accession sources. This allows us to add "data_source" to an
object that already has accession and source elements.